employee_tracker/requirements.txt
python-dotenv

fastapi
uvicorn
streamlit
